Agora os players já podem se cadastrar e esperar o início da partida

+++ src/main/java/com/fernando/carrotback/domain/model/PlayerAnswer.java
@@ -15,6 +15,6 @@ public class PlayerAnswer {
 
     private Long idQuestion;
     private Byte answer;
-    private Integer timeAnswerInSeconds;
+    private Long timeAnswerInSeconds;
     private Boolean correct;
 }
